backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.13.3
websockets==13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-generativeai==0.8.3
numpy==2.1.3
scikit-learn==1.5.2
